requirements.txt
gymnasium>=0.29
stable-baselines3>=2.3
torch>=2.0
numpy>=1.24
pyyaml>=6.0
matplotlib>=3.7
pandas>=2.0
scipy>=1.11
pytest>=7.4
pytest-cov>=4.1
ruff>=0.1
